openai-mcpkit/python-authenticated-mcp-server-scaffold/requirements.txt
mcp[server]>=1.0.0b6
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
openai>=1.40.0
pydantic>=2.7.0
PyJWT[crypto]>=2.8.0
httpx>=0.27.0
python-dotenv>=1.0.1
